+++ src/strings.ts
@@ -37,13 +37,15 @@ export const LIVE_SHARE_INVITE = name =>
   `${name} has invited you to a Live Share collaboration session.`;
 
 export const LIVE_SHARE_CHAT_NO_SESSION =
-  "Chat requires an active VS Live Share collaboration session.";
-
-export const LIVE_SHARE_CHAT_NOT_INITIATED =
-  "The Live Share session host has not initiated chat.";
+  "Chat requires an active Live Share collaboration session.";
 
 export const LIVE_SHARE_INFO_MESSAGES = {
   started: "_has started the Live Share session_",
   joined: "_has joined the Live Share session_",
   left: "_has left the Live Share session_"
 };
+
+export const LIVE_SHARE_CONFIRM_SIGN_OUT = provider =>
+  `To use chat over VS Live Share, you need to sign out of your ${provider} account.`;
+
+export const SIGN_OUT = "Sign out";